Move missed integration files to DIRT
