+++ test/Core.Test/Dirt/Utilities/IntegrationTemplateProcessorTests.cs
@@ -1,11 +1,11 @@
-﻿#nullable enable
+#nullable enable
 
-using Bit.Core.AdminConsole.Utilities;
+using Bit.Core.Dirt.Utilities;
 using Bit.Core.Models.Data;
 using Bit.Test.Common.AutoFixture.Attributes;
 using Xunit;
 
-namespace Bit.Core.Test.AdminConsole.Utilities;
+namespace Bit.Core.Test.Dirt.Utilities;
 
 public class IntegrationTemplateProcessorTests
 {
